requirements.txt
markdown
pyyaml
staticjinja
